refactor: moved the post processing functions into a seperate file
I created a sub package for the postInspector endpoint to keep the code a bit cleaner and to make it easier to find content in the future.
