+++ packages/backend/api/postInspector/postInspector.go
@@ -1,4 +1,4 @@
-package api
+package postInspector
 
 import (
 	"bytes"
@@ -6,10 +6,10 @@ import (
 	"fmt"
 	"net/http"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/FreiFahren/backend/Rstats"
+	"github.com/FreiFahren/backend/api"
 	"github.com/FreiFahren/backend/data"
 	"github.com/FreiFahren/backend/database"
 	_ "github.com/FreiFahren/backend/docs"
@@ -121,7 +121,7 @@ func processRequestData(req structs.InspectorRequest) (*structs.ResponseData, *s
 	}
 
 	if req.StationName != "" {
-		if stationID, found := FindStationId(req.StationName, stations); found {
+		if stationID, found := api.FindStationId(req.StationName, stations); found {
 			pointers.StationNamePtr = &req.StationName
 			pointers.StationIDPtr = &stationID
 			response.Station = structs.Station{Name: req.StationName, ID: stationID}
@@ -133,7 +133,7 @@ func processRequestData(req structs.InspectorRequest) (*structs.ResponseData, *s
 	}
 
 	if req.DirectionName != "" {
-		if directionID, found := FindStationId(req.DirectionName, stations); found {
+		if directionID, found := api.FindStationId(req.DirectionName, stations); found {
 			pointers.DirectionNamePtr = &req.DirectionName
 			pointers.DirectionIDPtr = &directionID
 			response.Direction = structs.Station{Name: req.DirectionName, ID: directionID, Coordinates: structs.Coordinates(stations[directionID].Coordinates), Lines: stations[directionID].Lines}
@@ -158,119 +158,6 @@ func processRequestData(req structs.InspectorRequest) (*structs.ResponseData, *s
 	logger.Log.Info().Interface("Response", response).Msg("Response data")
 
 	return response, pointers, nil
-}
-
-func postProcessInspectorData(dataToInsert *structs.ResponseData, pointers *structs.InsertPointers) error {
-	logger.Log.Debug().Msg("Filling missing columns using provided data")
-
-	var stations = data.GetStationsList()
-	var lines = data.GetLinesList()
-
-	if dataToInsert.Line == "" && dataToInsert.Station.ID != "" {
-		if err := assignLineIfSingleOption(dataToInsert, pointers, stations[dataToInsert.Station.ID]); err != nil {
-			logger.Log.Error().Err(err).Msg("Error assigning line if single option in postInspector")
-			return err
-		}
-	}
-
-	if dataToInsert.Direction.ID == "" && dataToInsert.Line != "" && dataToInsert.Station.ID != "" {
-		if err := determineDirectionIfImplied(dataToInsert, pointers, lines[dataToInsert.Line], dataToInsert.Station.ID); err != nil {
-			logger.Log.Error().Err(err).Msg("Error determining direction if implied in postInspector")
-			return err
-		}
-	}
-
-	// If the direction is the same as the station, remove the direction
-	if dataToInsert.Direction.ID == dataToInsert.Station.ID {
-		dataToInsert.Direction = structs.Station{}
-		pointers.DirectionIDPtr = nil
-		pointers.DirectionNamePtr = nil
-		logger.Log.Debug().Msg("Removed direction because it was the same as the station")
-	}
-
-	// if Station is not on the given line, remove the line
-	if dataToInsert.Station.ID != "" && !structs.StringInSlice(dataToInsert.Line, dataToInsert.Station.Lines) {
-		dataToInsert.Line = ""
-		pointers.LinePtr = nil
-		logger.Log.Debug().Msg("Removed line because the station was not on it")
-	}
-
-	// If Direction is not on the same given line, remove the direction
-	if dataToInsert.Direction.ID != "" && !structs.StringInSlice(dataToInsert.Line, dataToInsert.Direction.Lines) {
-		dataToInsert.Direction = structs.Station{}
-		pointers.DirectionIDPtr = nil
-		pointers.DirectionNamePtr = nil
-		logger.Log.Debug().Msg("Removed direction because it was not on the same line")
-	}
-
-	return nil
-}
-
-func assignLineIfSingleOption(dataToInsert *structs.ResponseData, pointers *structs.InsertPointers, station structs.StationListEntry) error {
-	logger.Log.Debug().Msg("Assigning line if single option")
-
-	// If there is only one line for the station, assign it
-	if len(station.Lines) == 1 {
-		dataToInsert.Line = station.Lines[0]
-		pointers.LinePtr = &dataToInsert.Line
-	}
-
-	// if there is only one line for the direction, assign it
-	if len(dataToInsert.Direction.Lines) == 1 {
-		dataToInsert.Line = dataToInsert.Direction.Lines[0]
-		pointers.LinePtr = &dataToInsert.Line
-	}
-
-	return nil
-}
-
-func determineDirectionIfImplied(dataToInsert *structs.ResponseData, pointers *structs.InsertPointers, line []string, stationID string) error {
-	logger.Log.Debug().Msg("Determining direction if implied")
-
-	var stations = data.GetStationsList()
-	isStationUniqueToOneLine := checkIfStationIsUniqueToOneLineOfType(stations[stationID], dataToInsert.Line)
-
-	lastStationID := line[len(line)-1]
-	firstStationID := line[0]
-
-	if isStationUniqueToOneLine {
-		if firstStationID == stationID {
-			if lastStation, found := stations[lastStationID]; found {
-				setDirection(dataToInsert, pointers, lastStationID, lastStation)
-			}
-		} else if lastStationID == stationID {
-			if firstStation, found := stations[firstStationID]; found {
-				setDirection(dataToInsert, pointers, firstStationID, firstStation)
-			}
-		}
-	}
-	return nil
-}
-
-// checks if a station is uniquely served by one line of the specified type (e.g., 'S' or 'U').
-func checkIfStationIsUniqueToOneLineOfType(station structs.StationListEntry, line string) bool {
-	logger.Log.Debug().Msg("Checking if station is unique to one line of the specified type")
-
-	// The first character of the line determines if it is a sbahn or ubahn
-	linePrefix := line[0]
-
-	count := 0
-	for _, line := range station.Lines {
-		if strings.HasPrefix(line, string(linePrefix)) {
-			count++
-		}
-	}
-
-	return count == 1 // return true if there is only one line of the specified type
-}
-
-func setDirection(dataToInsert *structs.ResponseData, pointers *structs.InsertPointers, stationID string, station structs.StationListEntry) {
-	logger.Log.Debug().Msg("Setting direction")
-
-	dataToInsert.Direction = structs.Station{Name: station.Name, ID: stationID, Coordinates: structs.Coordinates(station.Coordinates)}
-	pointers.DirectionIDPtr = &stationID
-	directionName := station.Name
-	pointers.DirectionNamePtr = &directionName
 }
 
 func notifyTelegramBotAboutReport(data *structs.ResponseData) error {
